Add Ansible version to a job's Zuul vars
The Ansible version is sometimes used for selecting the correct linter
or for implementing feature switches to make roles/playbooks backward
compatible.

With the split of Ansible into an "ansible" and "ansible-core" package,
the `ansible_version` now contains the version of the core package.
There seems to be no other variable that contains the version of the
"Ansible community" package that Zuul is using.

In order to support this use-case for Ansible 5+ we will add the Ansible
version to the job's Zuul vars.

Change-Id: I3f3a3237b8649770a9b7ff488e501a97b646a4c4