web: console: convert to PF4 DataList
The console page is currently based on the ListView from PF3.  The
most direct conversion to PF4 is a DataList, which is similar but
different.

This is intended to be a like-for-like replacement of the console page
using the DataList model.

The things I have explicitly thought about:

 * The trusted flag on the playbook remains; it now has a popup tip.
   For symmetry, untrusted playbooks are also marked.

 * The task line output is the same order, with

    task  <+>-  (STATUS)  [hosts]  <time>

   I've put these in separate data cells so that when it folds in on
   smaller devices the breakpoints can work however PF4 sets it up.

 * You can click on either the magnifying glass or the status to bring
   up the modal with the full task status.

 * The status and hosts are modified to use PF4 label and chip models
   for a standard look.

 * Plays start rolled up, but tasks start in the expanded state as
   before.

 * If there is a failed task, the playbook and play are unrolled to
   show it automatically as before.

 * Tasks highlight on hover, but now using a light boxshadow rather
   than background color which is more consistent with other parts of
   the ui (like the build table, for example).

 * No colours in the background of the playbook rows; the DataList now
   has a blue line that runs down the side showing you the groupings
   when a playbook row is opened.  I think this is more consistent
   with "less is more" type approach.

While I think there is plenty of room for improvement in the way this
information is displayed, I've deliberately tried to keep everything
the same in this changeset to a) ease review and b) so we have a
PF4-based grounding to work from.  There wasn't really a way to do
this more incrementally; although almost everything moves around there
is no tricky code to call out -- some fiddling of things that needed
to state properties and some toggle javascript-code bits are the main
additions.

Change-Id: Ie480deb046502879542e41844e919a362203e25d