Tracing: implement span save/restore
This adds API methods to the Tracing class to allow us to save and
restore spans using ZooKeeper data.

Additionally, we subclass the tracing.Span class so that we can
transparently handle timestamps which are stored as floating point
numbers rather than integer nanoseconds.

To exercise the new features, emit spans for QueueItems and BuildSets.

Change-Id: If55b06572b5e95f8c21611b2a3c23f7fd224a547