Fix and improve Keycloak tutorial
The keycloak tutorial incorrectly instructed users to run
"docker-compose-compose".  Correct that.

Also, change the instructions to "stop" rather than "down" the
original containers so that the results of the quick-start tutorial
are still present.

Finally, verify that, and also add a verification that the intended
effect of the restart worked (by checking the available authn methods).

Change-Id: I43a17e27300126e8acdc1919ba2bbe98719ad604