web: console: compress datalist
This is a selection of rules to override the default datalist layout;
it removes the lines and compresses entries.  This is the result of
inspection and trial-and-error; i.e. there may be better ways to do
all of this.

The box-shadow on hover mouseover doesn't really work that well with
the compressed entries, this replaces it with a light background.
Perhaps we don't need this at all?

Change-Id: Ie3d79b11be517136a62db61c57a112885b0e10eb