web: task summary: make more consistent with other tabs
After converting the console to PF4 with
Ie480deb046502879542e41844e919a362203e25d, flipping back through the
tabs I notied some inconsistency with the "Task Summary" tab.

Firstly this puts a <br> so it is vertically aligned the same as the
other tabs.  We remove the card around it because nothing else uses
this; now it is a DataList that is the same as the console page.  CSS
is added to remove the top line which is also inconsistent with the
other tabs.

We convert the node to a chip with a node icon as used elsewhere.

Future work would probably be to put the failed tasks as expandable
content in each node's summary line.

Change-Id: Ic134cf560cc78540f6af71964ac366f89ca594d9