web: console: better task modal titles
The task modal title is very sparse; since this can be deep-linked it
should provide more context.

This adds the task description as the main title, then adds the
status, node and deeplink icon as the modal-description.  This should
give more context to what is going on, and everything is styled the
same as on the console page so it's visually consistent.

Change-Id: I804d2eba9d67ed3a8781aa15ab6a14a2434329bb