Use json view for complex artifact metadata
If a job returns nested metadata (ie, a value that is a dict or
list), the build page will crash when attempting to display this
because it expects metadata values to only be scalars.  The docs
and code both say that values can be "anything".  Update the
build page so that if the metadata value is an object (ie, not a
primitive scalar), we use a JSON view to display it.

Also, if the value is not an object, use toString so that we
will display boolean values.

Change-Id: I609f70ec60399c69b5f8715b97c43f513c7360a1