feat(elasticsearch): support datastreams and rollover aliases
Add an option to disable index creation and disable prefixing the index name.
When using datastreams, Elasticsearch creates the backing indices automatically.
Set `_op_type` to create as it is required to send docs in a datastream and won't change
the current behavior as the reporter doesn't generate document IDs[1].

[1] https://www.elastic.co/guide/en/elasticsearch/reference/current/docs-index_.html#docs-index-api-query-params

Change-Id: Id783f34c53c87b20e051048cacb5975155e07360