Add semaphores to REST API
This adds information about semaphores to the REST API.

It allows for inspection of the known semaphores in a tenant, the
current number of jobs holding the semaphore, and information about
each holder iff that holder is in the current tenant.

Followup changes will add zuul-client and zuul-web support for the
API, along with docs and release notes.

Change-Id: I6ff57ca8db11add2429eefcc8b560abc9c074f4a