web: fix logo CSS
This was suggsted and written by Joachim Schuler in [1].  It is
probably a bug that due to asynchronous loading of CSS rules our
logo-height rule here sometimes comes and goes.  However, putting this
in a zuul-page class both seems to fix this (I always see the logo
now) and is just generally a nicer way to organise things.

[1] https://github.com/patternfly/patternfly-react/issues/7960#issuecomment-1244017295

Change-Id: Ib1b6be1b554b30ce7a7d2aacae476cefbed63782