web: update whitespace around logo
The current logo has a lot of whitespace around it and is offset with
the viewbox.  I've imported this into inkscape, selected just the logo
and re-exported it with no whitespace, then ran it through the
minimizer at [1].

I've set the height to 1.5em -- the icons in the same bar have a
height of 1em; to my eye this look good just a bit bigger than that.

[1] https://jakearchibald.github.io/svgomg/

Change-Id: Idf4e676d33a4626c0b80e06986d8b24b41924065