web: Even "better" hidden expandable content
To recap, a PF DataList doesn't have a built-in way to hide the
"expandable content" toggle, so on the console page the items -- which
have a mixture of expandable content and not -- don't line up.

In I33fdafb4e5b98e546c8615a3d65f52c944521cf7 I tried to fix this by
using a visibility:hidden toggle, based on ideas from the comments in
the inline github issue.

What I didn't realise is this doesn't work quite correctly on very
narrow (mobile) views.  The problem is that the class hierarchy needs
to look like

 <row>
   <control>
    <toggle button />
   </control>
   <content>
    <cell />
    <cell />
     ...
   </content>
 </row>

The reason for this is that the <content> is a grid placed *next* to
the <control>.  So when it squishes up, it should look like

 |-- |
 | > | name
 |   | <magnifying> <status>
 |   | <time>
 |---|

With my first attempt, the hidden toggle ">" had all the right
classes, but was part of the content.  This looks fine when all the
content is in a horizontal line; but when it gets to a very small page
width we end up with a hidden toggle button in a single column *above*
the name, pushing everything down

 | > hidden button <
 | name
 | <magnifying> <status>
 | <time>

This leads to a large gap between items and misalignment on small
width devices.

Upon further reflection, it has struck me that we don't *have* to use
the React compontents, and what we need to do here is build our own
datalist item from the PF classes directly, as if we were writing with
PF in HTML.  We do this here, adding the hidden to the expandable
button.

This works on all my testing, and the generated output via the
browsers console inspector looks the same as using the React
components.  Obviously if PF merge a change where we can do this via a
prop to the compontent it is much more idiomatic to React code, but
this should work fine too.

Change-Id: I2048554f46b6ef3ec5410967e00400a593c572d7