web: Differentiate console entries
I've noticed that when making the console page very narrow (mobile)
the entries are difficult to differentiate in their compressed form.

This proposes adding alternating colors on the play entries to
differentiate them.  Since the grey is used between elements, the
highlight is modified to the the lightest blue in the PF family.
Putting a small border-radius on this makes it look a bit more
differentiated.

Change-Id: I7afcb64a5e41348f14bd297c8c8bde27312dff97