web: Simply task status results
Currently this displays the task OK, Changed and Failure(s) results in
a traffic-light format.  I've been looking at the way some other
similar tools present this sort of info, and I think that a few things
stand out

 - the orange for the "changed" status is a bit of a red-herring,
   because it suggests a warning but really a changed task is as
   "good" as an OK task.

 - Secondly, the failure case doesn't stand out enough; failures are
   going to be the cause of problems, but you get the same
   traffic-light for hosts that ran successfully and failed.

 - When nothing has failed, we show "0 failures" -- but that is pretty
   much implied by things working.  It's a bit redundant and puts up a
   red box when there isn't a failure.

This proposes moving into a single tasks results label that is either
green for success, or red if there are failures.  If there are no
failures, you just see the count of OK/Changed tasks -- if there are,
you'll also see the failed count.

Change-Id: Iebadc4ddb77209f69242ebc5ac46f2ae6d67789f