web: Put status icon into one label
In some recent changes I've been looking at the narrow view of the
console page.  One thing that remains inconsistent is the magnifying
arrow; it can jump around depending on the page width and in my
subjective opinion it looks a little out of context.

This proposes making a single "pill" for the status and magnifying
icon.  The width should be consistent, and the magnifying glass
floating right behind a divider to inidicate it is separate to the
status -- it is also set to not break so it stays together.

Change-Id: I183d4c939a1c74e21198564d1d51e33ecdbc813c