Create link to previous buildset span
Create a link to the previous buildset span on gate reset. To make this
work we'll start the buildset span when the buildset is created instead
of only when we set the configuration.

This change also adds the `is_remote` flag of the span context of
related links. This is required for creating a `SpanContext` in order to
deserialize the links.

Change-Id: If3a3a83739c1472659d71d05dcf67f84ddce4247