Add option to provide detailed event build param
Currently the build parameters provide just the 'event_id' mostly for
debugging purposes. However, providing the full event details can open
up the ability for more customization of jobs run on a per-event basis.

As an example for the gerrit driver, commenting just 'recheck' could run
regular 'check' pipeline jobs, but 'recheck tests' could run a variant
of such jobs instead. But to be able to do so, the instigating event
information needs to be provided.

As such, this patch adds a new configuration option under the 'executor'
section named 'detailed_event' which, if set to true, would print out a
dict of the event data as part of the build parameters.

The default is false, keeping the current behaviour of just providing
the event_id.

Change-Id: Ic3a3bc9276f7e0bea45d9ce22fc27943e4c93ce3