Update patch set 1
Patch Set 1: Code-Review-1

We intentionally don't provide this information because there are many ways for changes to be enqueued into pipelines and we want to maintain an equivalency between them.  The pipeline itself defines the workflow action, not the job.

That isn't to say that we're not open to the idea of including more information, however, consider that a paramount use case is being able to enqueue changes manually using zuul-client and having them run exactly the same way as if someone had just uploaded a new patchset.  Once a change is in a pipeline, regardless of how it gits there, what happens next is deterministic.  A change like this would not permit that.

Zuul is an abstraction layer that smooths over differences between its drivers and triggers, so a solution needs to account for that and be universal.

Patch-set: 1
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1