Link span of queue item to trigger event span
This change links the span of the queue item to the span of the
forwarded trigger event.

This adds a currently empty span context field to trigger events which
will be later populated by the drivers.

We also need to add add a span context field to
`ChangeManagementEvent`s, as some drivers add dequeue events to the
pipeline event queues directly.

Change-Id: Icd240712b86cc22e55fb67f6787a0974d5308043