Trace received Github events
We'll create a span when zuul-web receives a Github webhook event which
is then linked to the span for the event pre-processing step.

The pre-processing span context will be added to the trigger events and
with Icd240712b86cc22e55fb67f6787a0974d5308043 complete tracing of the
whole chain from receiving a Github event until a change is enqueued.

Change-Id: I1734a3a9e44f0ae01f5ed3453f8218945c90db58