Simplify tenant_authorizatons check
This method iterates over all tenants but only needs to return
information about a single tenant.  Simplify the calculation for
efficiency.

This includes a change in behavior for unknown tenants.  Currently,
a request to /api/tenant/{name}/authorizations will always succeed
even if the tenant does not exist (it will return an authorization
entry indicating the user is not an admin of the unknown tenant).
This is unnecessary and confusing.  It will now return a 404 for
the unknown tenant.

In the updated unit test, tenant-two was an unknown tenant; its name
has been updated to 'unknown' to make that clear.

(Since the test asserted that data were returned either way, it is
unclear whether the original author of the unit test expected
tenant-two to be unknown or known.)

Change-Id: I545575fb73ef555b34c207f8a5f2e70935c049aa