Add check_auth tool to zuul-web
Authentication checking in the admin methods of zuul-web is very
duplicative.  Consolidate all of the auth checks into a cherrypy
tool that we can use to decorate methods.

This tool also anticipates that we will have read-only checks in
the future, but for now, it is still only used for admin checks.

This tool also populates some additional parameters (like tenant
and auth info) so that we don't need to call "getTenantOrRaise"
multiple times in a request.

Several methods performed HTTP method checks inside the method
which inhibits our ability to wrap an entire method with an
auth_check.  To resolve this, we now use method conditions on
the routes dispatcher.  As a convention, I have put the
options handling on the "GET" methods since they are most likely
to be universal.

Change-Id: Id815efd9337cbed621509bb0f914bdb552379bc7