Web: always set redux auth and wait for it
Currently we only set the auth redux property within the tenant
scope.  But we're going to extend authentication to outside the
tenant scope (to optionally restrict access to the tenant list).
To accommodate that, we will always update the auth property in
redux any time our tenant changes (including when it changes to
undefined), and we will wait until we have finished auth configuration
before rendering any pages in the app.

This is effectively a no-op change from the end-user point of view,
except that we may wait just a bit longer (for the /api/tenant/info
endpoint to return) before showing an initial page for a tenant.

Change-Id: I18e74fa205f75a7b020bf23c8652226e5170d88b