Set default SSH_AUTH_SOCK in zuul-bwrap command
The zuul-bwrap command is useful for debugging things under the zuul
bwrap environment. Unfortunately, the way things are written it assumes
there will be an SSH_AUTH_SOCK. For much debugging you might manually do
in this environment an SSH_AUTH_SOCK is unnecessary. Instead of throwing
a obtuse error simply set the value to /dev/null if not otherwise set.

Change-Id: Iec0ee93c6e6b1b647a27c9a7fdf280d14d5d2596