configloader: Fix error report for job nodesets
In a job nodeset, you do not need a "name", but some of the error
logic assumes it.

This leads to a rather confusing error return; for example something
like

 job:
  ...
  nodeset:
    nodes: []
    groups:
      - name: a_group
        nodes:
          - a_node_that_does_not_exist

leads to

 Zuul encountered a syntax error while parsing its configuration in the
 repo opendev/system-config on branch master.  The error was:

  'name'

when really it is trying to tell you is that "this_node_not_defined"
isn't defined, but it tries to de-reference the name of the nodeset:
and fails.

This keeps track of the anonymous nodeset flag and if set, then
rewords the error to make sense in the job context; for example it
will now say:

 Zuul encountered a syntax error while parsing its configuration in the
 repo org/project on branch master.  The error was:

   In the nodeset the group "a_group" contains a node named
   "a_node_that_does_not_exist" which is not defined in the nodeset.

  The error appears in the following job stanza:

    job:
        name: test job
        nodeset:
          nodes: []
          groups:
            - name: a_group
              nodes:
               - a_node_that_does_not_exist

Test cases are added to cover the updated messages.

Change-Id: Ia24e66ba9b97618636023cb61d7c82aa7306d678