Don't trace merge jobs that we don't lock
We get a trace from every merger (including executors) for every
merge job because we start the trace before attempting the lock.
So essentially, we get one trace from the merger that runs the job,
and one trace from every other merger indicating that it did not
run the job.

This is perhaps too much detail for us.  While it's true that we
can see the response times of every system component here, it may
be sufficient to have only the response time of the first merger.
This will reduce the noise in trace visualizations significantly.

Change-Id: I88c56f00c060eae9316473f4a4e222a0db97e510